refactor(TicketReply): extract to seperate file
